+++ src/core/utils/common.ts
@@ -135,10 +135,6 @@ export function deepClone(obj): any {
     : clone
 }
 
-export function mergeObjects(a: any, b: any) {
-  return Object.assign({}, a, b)
-}
-
 export function sortObject(unordered) {
   return Object.keys(unordered)
     .sort()
@@ -146,4 +142,31 @@ export function sortObject(unordered) {
       obj[key] = unordered[key]
       return obj
     }, {})
+}
+
+export function serializeFunc(fn) {
+  //the source code from https://github.com/yahoo/serialize-javascript/blob/main/index.js
+  const serializedFn = fn.toString()
+  if (/function.*?\(/.test(serializedFn)) {
+    return serializedFn
+  }
+  if (/.*?=>.*?/.test(serializedFn)) {
+    return serializedFn
+  }
+  const argsStartsAt = serializedFn.indexOf('(')
+  const def = serializedFn
+    .substr(0, argsStartsAt)
+    .trim()
+    .split(' ')
+    .filter((val) => val.length > 0)
+  const nonReservedSymbols = def.filter((val) => ['*', 'async'].indexOf(val) === -1)
+  if (nonReservedSymbols.length > 0) {
+    return (
+      (def.indexOf('async') > -1 ? 'async ' : '') +
+      'function' +
+      (def.join('').indexOf('*') > -1 ? '*' : '') +
+      serializedFn.substr(argsStartsAt)
+    )
+  }
+  return serializedFn
 }